requirements.txt
openai>=1.0.0
streamlit>=1.22.0
pillow>=9.0.0
python-dotenv>=0.19.0
requests>=2.28.0
pydantic>=1.9.0
reportlab>=4.0.0          # For PDF report generation
email-validator>=2.0.0    # For validating notification/CRM email fields
typing-extensions>=4.0.0  # Ensures compatibility with type hints on older Python
pytest>=7.0.0             # For running tests (optional, but recommended)
watchdog>=3.0.0           # For development hot-reloading (optional)